Chance of obtaining fish from fishing and the chance of receiving a skillup when successfully catching a fish have been reverted to 1.12 chances. The speed in which you catch fish has been kept in its 3.3.5 state. (#20)
internal note: It's my belief that TrinityCore actually has a better implementation than Vmangos, so its been kept mostly intact. TC uses a system called "fishingsteps" which is the amount of times you have to catch a fish before your next skillup. Vmangos uses pure chance. It's well theorized by many sources, such as El's Anglin', that it's likely WoW uses a Fishing Steps-like system. However it's supposed to be an average number of catches, not a set number. Due to that I changed it so that we still use fishingsteps but when you reach the required number of steps each cast now has a 75% chance of skilling you up. It's not perfectly blizzlike, but its close enough.

Fishing from a pool of fish no longer guarantees a catch regardless of your fishing skill, as that was a change made in patch 3.3.0.

Added new hook for checking your fishing cast, allowing you to make a fishing cast fail. Includes the liquidtype that you're attempting to fish in.

Added LiquidTypes enum