Rework stealth detection formula to mimic Vmangos, bringing it much closer to Classic. (#22)
Fixed a bug where stealth effects could never be seen through if the stealther was behind the target, even if that target was extremely higher level than you.

Compared results with Classic to ensure accuracy, unfortunately I do not have a rogue so I was only able to test using a fresh level 1, whom I leveled to 2 as well. However I was able to replicate perfect accuracy with the limited sample size.