Add range/alive/LoS checks for casts with implicit pet target.
If there are no checks, spells like mend pet fail silently if the pet is
out of range / dead / out of LoS, as if the pet "evades" the spell (but
mana is still taken). This should solve most of the issues one might
encounter.

Code is taken from VMangos -- apart from the LoS ignoring M2 which seems
to be common throughout TC code.

To investigate: Improved Mend Pet fails LoS Check. (unaffected by this
commit)