Do not allow non-participants to interact with corpse/gameobject at all.
This solves a few edge-cases where if someone entered a dungeon with a
dead boss, they would be able to loot with FFA (or elapsed group loot
timer) the boss loot. The person would become able to interact with the
corpse as soon as anyone who was currently looting would close the loot
window.