Rework Attack Power Calculations to use positive and negative mods.
Based on vmangos commit at https://github.com/vmangos/core/commit/2646715596d24ba205b03a7f3e1cee1c4ffa4ab9

There are two AP modifiers that the client expects, a positive one and a
negative one. TC was not handling them separately in some cases (in
attPowerMod), which led to bugs during applications and deapplications
of AP auras.

Most of the magic is inside UpdateAttackPowerAndDamage, the AP modifiers
(+, - and multiplier) are kept inside m_attackPowerMods (which replace
the AP m_auraFlatModifiersGroup + pct), which is being updated via AuraEffects.

Store the base AP of creatures so that we do not fetch them on every
UpdateAttackAndPower.