Fix hunter traps being invisible.
When calculating the detect skill of the player vs the trap stealth,
the level of the trap was included twice in the calculations, rendering
the traps virtually invisible unless the player massively overleveled
the hunter.

Default trap stealth value is 70, which corresponds to a 14 level
difference, rendering it visible for rogues but still invisible for other
classes.